Create custom Prisma adapter for Next auth
This new adapter enables us to make use of the `Profile` model. When a
user is created, a random hex string of max username length (or one
less) is assigned to the user's profile as their username. This is
because `username` is a required field, and I don't quite know how to
integrate custom data + provider data together while creating an
account; therefore, it will be up to the user to change their username
once their account has been created.

Also export min and max values more thorougly and in a different style,
along with creating a username schema.

The `email` field of the `User` model is now required. I did this
because the `AdapterUser` type that Next auth adapters use requires that
the email field strictly be a `string` and not nullish. Funnily enough,
the included Prisma adapter does not follow this rule.